# pos_app/docs/openapi.yaml
openapi: 3.0.3
info:
  title: POS Multi-Sucursal API
  version: 1.0.0
  description: Punto de venta y control de inventario multi-sucursal.
paths:
...
  /api/v1/inventory/:
    get:
      operationId: inventory_list
      description: Existencias de la sucursal.
      parameters:
      - in: query
        name: branch
        schema:
          type: string
        description: Código de la sucursal. Por defecto, la asignada al usuario.
      - name: page
        required: false
        in: query
        description: Un número de página dentro del conjunto de resultados paginado.
        schema:
          type: integer
      - name: page_size
        required: false
        in: query
        description: Número de resultados a devolver por página.
        schema:
          type: integer
      tags:
      - inventory
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedBranchInventoryList'
          description: ''
  /api/v1/inventory/{product_id}/minimum-stock/:
    patch:
      operationId: inventory_set_minimum_stock
      description: Define el stock mínimo del producto en la sucursal.
      parameters:
      - in: path
        name: product_id
        schema:
          type: integer
        required: true
      tags:
      - inventory
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedMinimumStock'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedMinimumStock'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedMinimumStock'
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BranchInventory'
          description: ''
  /api/v1/inventory/low-stock/:
    get:
      operationId: inventory_low_stock
      description: Productos activos con stock en o por debajo del mínimo.
      parameters:
      - in: query
        name: branch
        schema:
          type: string
        description: Código de la sucursal. Por defecto, la asignada al usuario.
      - name: page
        required: false
        in: query
        description: Un número de página dentro del conjunto de resultados paginado.
        schema:
          type: integer
      - name: page_size
        required: false
        in: query
        description: Número de resultados a devolver por página.
        schema:
          type: integer
      tags:
      - inventory
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedBranchInventoryList'
          description: ''
  /api/v1/inventory/movements/:
    get:
      operationId: inventory_movements_list
      description: Kardex de la sucursal, del movimiento más reciente al más antiguo.
      parameters:
      - in: query
        name: branch
        schema:
          type: string
      - in: query
        name: created_from
        schema:
          type: string
          format: date-time
      - in: query
        name: created_to
        schema:
          type: string
          format: date-time
      - in: query
        name: movement_type
        schema:
          type: string
          enum:
          - ADJUSTMENT
          - ENTRY
          - SALE
          - WASTE
        description: |-
          * `ENTRY` - Entrada
          * `SALE` - Venta
          * `WASTE` - Merma
          * `ADJUSTMENT` - Ajuste
      - name: page
        required: false
        in: query
        description: Un número de página dentro del conjunto de resultados paginado.
        schema:
          type: integer
      - name: page_size
        required: false
        in: query
        description: Número de resultados a devolver por página.
        schema:
          type: integer
      - in: query
        name: product
        schema:
          type: number
      tags:
      - inventory
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedInventoryMovementList'
          description: ''
    post:
      operationId: inventory_movements_create
      description: |-
        Registra una entrada, merma o ajuste.

        Responde 204 si un ajuste coincide con el stock actual y no cambia nada.
      tags:
      - inventory
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InventoryMovementCreate'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/InventoryMovementCreate'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/InventoryMovementCreate'
        required: true
      security:
      - jwtAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InventoryMovement'
          description: ''
        '204':
          description: No response body
  /api/v1/products/:
    get:
      operationId: products_list
      description: Catálogo de productos, ordenado por nombre.
      parameters:
      - in: query
        name: active
        schema:
          type: boolean
        description: Solo productos activos.
      - in: query
        name: category
        schema:
          type: string
          enum:
          - ACCESSORIES
          - LIQUIDS
          - POWDERS
      - name: page
        required: false
        in: query
        description: Un número de página dentro del conjunto de resultados paginado.
        schema:
          type: integer
      - name: page_size
        required: false
        in: query
        description: Número de resultados a devolver por página.
        schema:
          type: integer
      - in: query
        name: search
        schema:
          type: string
        description: Texto dentro del nombre.
      tags:
      - products
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedProductList'
          description: ''
    post:
      operationId: products_create
      description: Crea un producto con stock cero en todas las sucursales activas.
      tags:
      - products
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Product'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Product'
        required: true
      security:
      - jwtAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
          description: ''
  /api/v1/products/{id}/:
    get:
      operationId: products_retrieve
      description: Lectura para SUPERVISOR y MANAGER; escritura solo para MANAGER.
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      tags:
      - products
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
          description: ''
    patch:
      operationId: products_partial_update
      description: Actualiza los campos enviados. No afecta a ventas ya facturadas.
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      tags:
      - products
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedProduct'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedProduct'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedProduct'
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
          description: ''
  /api/v1/products/{id}/toggle-active/:
    post:
      operationId: products_toggle_active
      description: Activa o desactiva el producto.
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      tags:
      - products
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
          description: ''
...
components:
  schemas:
...
    BranchInventory:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        product:
          type: integer
          readOnly: true
        product_name:
          type: string
          readOnly: true
        branch:
          type: string
          readOnly: true
        current_stock:
          type: string
          format: decimal
          pattern: ^-?\d{0,9}(?:\.\d{0,3})?$
          readOnly: true
        minimum_stock:
          type: string
          format: decimal
          pattern: ^-?\d{0,9}(?:\.\d{0,3})?$
          readOnly: true
      required:
      - branch
      - current_stock
      - id
      - minimum_stock
      - product
      - product_name
...
    CategoryEnum:
      enum:
      - LIQUIDS
      - POWDERS
      - ACCESSORIES
      type: string
      description: |-
        * `LIQUIDS` - Líquidos
        * `POWDERS` - Polvos
        * `ACCESSORIES` - Accesorios
...
    InventoryMovement:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        product:
          type: integer
          readOnly: true
        branch:
          type: string
          readOnly: true
        movement_type:
          allOf:
          - $ref: '#/components/schemas/InventoryMovementMovementTypeEnum'
          readOnly: true
        quantity:
          type: string
          format: decimal
          pattern: ^-?\d{0,9}(?:\.\d{0,3})?$
          readOnly: true
        stock_before:
          type: string
          format: decimal
          pattern: ^-?\d{0,9}(?:\.\d{0,3})?$
          readOnly: true
        stock_after:
          type: string
          format: decimal
          pattern: ^-?\d{0,9}(?:\.\d{0,3})?$
          readOnly: true
        user:
          type: integer
          readOnly: true
        sale:
          type: integer
          readOnly: true
          nullable: true
        notes:
          type: string
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - branch
      - created_at
      - id
      - movement_type
      - notes
      - product
      - quantity
      - sale
      - stock_after
      - stock_before
      - user
    InventoryMovementCreate:
      type: object
      description: Entrada, merma o ajuste manual. Las ventas generan su Kardex solas.
      properties:
        product_id:
          type: integer
          minimum: 1
        branch:
          type: string
          maxLength: 20
        movement_type:
          $ref: '#/components/schemas/InventoryMovementCreateMovementTypeEnum'
        quantity:
          type: string
          format: decimal
          pattern: ^-?\d{0,9}(?:\.\d{0,3})?$
          description: Cantidad que entra o se pierde. En un ADJUSTMENT es el stock
            contado.
        notes:
          type: string
          maxLength: 255
      required:
      - movement_type
      - product_id
      - quantity
    InventoryMovementCreateMovementTypeEnum:
      enum:
      - ENTRY
      - WASTE
      - ADJUSTMENT
      type: string
      description: |-
        * `ENTRY` - ENTRY
        * `WASTE` - WASTE
        * `ADJUSTMENT` - ADJUSTMENT
    InventoryMovementMovementTypeEnum:
      enum:
      - ENTRY
      - SALE
      - WASTE
      - ADJUSTMENT
      type: string
      description: |-
        * `ENTRY` - Entrada
        * `SALE` - Venta
        * `WASTE` - Merma
        * `ADJUSTMENT` - Ajuste
...
    PaginatedBranchInventoryList:
      type: object
      required:
      - count
      - results
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=4
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=2
        results:
          type: array
          items:
            $ref: '#/components/schemas/BranchInventory'
...
    PaginatedInventoryMovementList:
      type: object
      required:
      - count
      - results
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=4
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=2
        results:
          type: array
          items:
            $ref: '#/components/schemas/InventoryMovement'
    PaginatedProductList:
      type: object
      required:
      - count
      - results
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=4
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=2
        results:
          type: array
          items:
            $ref: '#/components/schemas/Product'
...
    PatchedMinimumStock:
      type: object
      properties:
        branch:
          type: string
          maxLength: 20
        minimum_stock:
          type: string
          format: decimal
          pattern: ^-?\d{0,9}(?:\.\d{0,3})?$
    PatchedProduct:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
          maxLength: 150
        category:
          $ref: '#/components/schemas/CategoryEnum'
        unit_of_measure:
          $ref: '#/components/schemas/UnitOfMeasureEnum'
        cost_price_usd:
          type: string
          format: decimal
          pattern: ^-?\d{0,12}(?:\.\d{0,2})?$
        sale_price_usd:
          type: string
          format: decimal
          pattern: ^-?\d{0,12}(?:\.\d{0,2})?$
        active:
          type: boolean
          readOnly: true
...
    Product:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
          maxLength: 150
        category:
          $ref: '#/components/schemas/CategoryEnum'
        unit_of_measure:
          $ref: '#/components/schemas/UnitOfMeasureEnum'
        cost_price_usd:
          type: string
          format: decimal
          pattern: ^-?\d{0,12}(?:\.\d{0,2})?$
        sale_price_usd:
          type: string
          format: decimal
          pattern: ^-?\d{0,12}(?:\.\d{0,2})?$
        active:
          type: boolean
          readOnly: true
      required:
      - active
      - category
      - cost_price_usd
      - id
      - name
      - sale_price_usd
      - unit_of_measure
...
    UnitOfMeasureEnum:
      enum:
      - LITER
      - KILOGRAM
      - UNIT
      type: string
      description: |-
        * `LITER` - Litro
        * `KILOGRAM` - Kilogramo
        * `UNIT` - Unidad
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT